# go_api/docs/swagger.yaml
definitions:
...
  dto.CreateUserRequest:
    properties:
      degree:
        type: string
      email:
        type: string
      employee:
        $ref: '#/definitions/dto.EmployeeProfileDTO'
      name:
        type: string
      phone_number:
        type: string
      send_login_credentials_email:
        type: boolean
      student:
        $ref: '#/definitions/dto.StudentProfileDTO'
      surname:
        type: string
    required:
    - email
    - name
    - surname
    type: object
...
  dto.EmployeeProfileDTO:
    properties:
      faculty_id:
        type: integer
      unit_id:
        type: integer
    type: object
...
  dto.StudentProfileDTO:
    properties:
      major_id:
        type: integer
      study_program_id:
        type: integer
    type: object
...
  models.PaginatedUsersResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
    type: object
...
  models.Unit:
    properties:
      faculty_id:
        type: integer
      id:
        type: integer
      unit_name:
        type: string
      unit_short:
        type: string
    type: object
  models.UnitResponse:
    properties:
      courses_count:
        type: integer
      faculty_id:
        type: integer
      id:
        type: integer
      lecturers_count:
        type: integer
      unit_name:
        type: string
      unit_short:
        type: string
    type: object
  models.UnitsListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/models.UnitResponse'
        type: array
    type: object
  models.UserDetailResponse:
    properties:
      created_at:
        type: string
      degree:
        type: string
      email:
        type: string
      employee:
        additionalProperties: true
        type: object
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      roles:
        items:
          type: string
        type: array
      student:
        additionalProperties: true
        type: object
      surname:
        type: string
      two_factor_enabled:
        type: boolean
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      degree:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
      roles:
        items:
          type: string
        type: array
      surname:
        type: string
      two_factor_enabled:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/units:
    get:
      description: Returns list of units with lecturers and courses counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UnitsListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get units
      tags:
      - units
    post:
      consumes:
      - application/json
      description: Creates a new unit.
      parameters:
      - description: Unit data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Unit'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Unit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create unit
      tags:
      - units
  /api/v1/users:
    get:
      description: Returns list of users with roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedUsersResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a user via internal gRPC service
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create user (proxy)
      tags:
      - users
  /api/v1/users/{id}:
    delete:
      description: Deletes a user via internal gRPC service
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user (proxy)
      tags:
      - users
    get:
      description: Fetches user details from internal gRPC service and enriches with
        DB data
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserDetailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID (proxy)
      tags:
      - users
swagger: "2.0"