# go_api/internal/handlers/v1/user_handler.go
package handlers

import (
	"context"
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/credentials/insecure"
	"google.golang.org/grpc/status"

	"go_api/db"
	"go_api/internal/dto"
	"go_api/internal/models"
	pb "go_api/internal/rpc/user"
)


// GetUsers godoc
// @Summary Get users
// @Description Returns list of users with roles
// @Tags users
// @Produce json
// @Success 200 {object} models.PaginatedUsersResponse
// @Failure 500 {object} map[string]string
// @Router /api/v1/users [get]
func GetUsers(c *gin.Context) {
	var users []models.User

	if err := db.DB.Preload("Roles").Order("id").Find(&users).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	var mappedItems []models.UserResponse
	for _, u := range users {
		roleNames := []string{}
		for _, r := range u.Roles {
			roleNames = append(roleNames, r.RoleName)
		}

		mappedItems = append(mappedItems, models.UserResponse{
			ID:               u.ID,
			Email:            u.Email,
			PhoneNumber:      u.PhoneNumber,
			Name:             u.Name,
			Surname:          u.Surname,
			Degree:           u.Degree,
			CreatedAt:        u.CreatedAt,
			Roles:            roleNames,
			TwoFactorEnabled: u.TwoFactorEnabled,
		})
	}

	c.JSON(http.StatusOK, models.PaginatedUsersResponse{
		Items: mappedItems,
	})
}


// GetUserProxy godoc
// @Summary Get user by ID (proxy)
// @Description Fetches user details from internal gRPC service and enriches with DB data
// @Tags users
// @Produce json
// @Param id path int true "User ID"
// @Success 200 {object} models.UserDetailResponse
// @Failure 400 {object} map[string]string
// @Failure 404 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/v1/users/{id} [get]
func GetUserProxy(c *gin.Context) {
	idParam := c.Param("id")
	userId, err := strconv.ParseInt(idParam, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid user ID format"})
		return
	}

	conn, err := grpc.Dial("backend:50051", grpc.WithTransportCredentials(insecure.NewCredentials()))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to connect to internal RPC service: " + err.Error()})
		return
	}
	defer conn.Close()

	client := pb.NewUserRpcServiceClient(conn)

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	resp, err := client.GetUserRPC(ctx, &pb.UserGetRequest{Id: int32(userId)})
	if err != nil {
		st, ok := status.FromError(err)
		if ok && st.Code() == codes.NotFound {
			c.JSON(http.StatusNotFound, gin.H{"error": st.Message()})
		} else {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal RPC error: " + err.Error()})
		}
		return
	}

	var dbUser models.User
	roleNames := []string{}
	if err := db.DB.Preload("Roles").First(&dbUser, userId).Error; err == nil {
		for _, r := range dbUser.Roles {
			roleNames = append(roleNames, r.RoleName)
		}
	}

	var phoneNumber *string
	if resp.PhoneNumber != "" {
		phoneNumber = &resp.PhoneNumber
	}
	var degree *string
	if resp.Degree != "" {
		degree = &resp.Degree
	}

	userDetail := models.UserDetailResponse{
		ID:               int(resp.Id),
		Email:            resp.Email,
		PhoneNumber:      phoneNumber,
		Name:             resp.Name,
		Surname:          resp.Surname,
		Degree:           degree,
		CreatedAt:        dbUser.CreatedAt,
		TwoFactorEnabled: dbUser.TwoFactorEnabled,
		Roles:            roleNames,
	}

	if resp.GetStudent() != nil {
		userDetail.Student = map[string]interface{}{
			"id":               resp.GetStudent().Id,
			"study_program_id": resp.GetStudent().StudyProgramId,
			"major_id":         resp.GetStudent().MajorId,
		}
	} else if resp.GetEmployee() != nil {
		userDetail.Employee = map[string]interface{}{
			"id":         resp.GetEmployee().Id,
			"faculty_id": resp.GetEmployee().FacultyId,
			"unit_id":    resp.GetEmployee().UnitId,
		}
	}

	c.JSON(http.StatusOK, userDetail)
}


// CreateUserProxy godoc
// @Summary Create user (proxy)
// @Description Creates a user via internal gRPC service
// @Tags users
// @Accept json
// @Produce json
// @Param request body dto.CreateUserRequest true "User data"
// @Success 201 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/v1/users [post]
func CreateUserProxy(c *gin.Context) {
	conn, err := grpc.Dial("backend:50051", grpc.WithTransportCredentials(insecure.NewCredentials()))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to connect to internal RPC service: " + err.Error()})
		return
	}
	defer conn.Close()

	client := pb.NewUserRpcServiceClient(conn)

	var reqBody dto.CreateUserRequest
	if err := c.ShouldBindJSON(&reqBody); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	grpcReq := &pb.UserCreateRequest{
		Email:                     reqBody.Email,
		Name:                      reqBody.Name,
		Surname:                   reqBody.Surname,
		PhoneNumber:               reqBody.PhoneNumber,
		Degree:                    reqBody.Degree,
		SendLoginCredentialsEmail: reqBody.SendLoginCredentialsEmail,
	}

	if reqBody.Student != nil {
		grpcReq.Profile = &pb.UserCreateRequest_Student{
			Student: &pb.StudentProfile{
				StudyProgramId: reqBody.Student.StudyProgramId,
				MajorId:        reqBody.Student.MajorId,
			},
		}
	} else if reqBody.Employee != nil {
		grpcReq.Profile = &pb.UserCreateRequest_Employee{
			Employee: &pb.EmployeeProfile{
				FacultyId: reqBody.Employee.FacultyId,
				UnitId:    reqBody.Employee.UnitId,
			},
		}
	}

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	resp, err := client.CreateUserRPC(ctx, grpcReq)
	if err != nil {
		st, ok := status.FromError(err)
		if ok {
			c.JSON(http.StatusBadRequest, gin.H{"error": st.Message()})
		} else {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal RPC error: " + err.Error()})
		}
		return
	}

	c.JSON(http.StatusCreated, gin.H{
		"id":     resp.Id,
		"email":  resp.Email,
		"status": resp.Status,
	})
}


// DeleteUserProxy godoc
// @Summary Delete user (proxy)
// @Description Deletes a user via internal gRPC service
// @Tags users
// @Produce json
// @Param id path int true "User ID"
// @Success 200 {object} map[string]interface{}
// @Failure 400 {object} map[string]string
// @Failure 500 {object} map[string]string
// @Router /api/v1/users/{id} [delete]
func DeleteUserProxy(c *gin.Context) {
	idParam := c.Param("id")
	userId, err := strconv.ParseInt(idParam, 10, 32)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid user ID format"})
		return
	}

	conn, err := grpc.Dial("backend:50051", grpc.WithTransportCredentials(insecure.NewCredentials()))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to connect to internal RPC service: " + err.Error()})
		return
	}
	defer conn.Close()

	client := pb.NewUserRpcServiceClient(conn)

	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
	defer cancel()

	resp, err := client.DeleteUserRPC(ctx, &pb.UserDeleteRequest{Id: int32(userId)})
	if err != nil {
		st, ok := status.FromError(err)
		if ok {
			c.JSON(http.StatusBadRequest, gin.H{"error": st.Message()})
		} else {
			c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		}
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"success": resp.Success,
		"message": resp.Message,
	})
}